Add files via upload
A cricket game that is console based and utilizes object oriented programming concepts. The user will first be asked for overs and then the toss. If the user selects bowling first, the computer will bat and vise versa. The user will have to specify the angle (like for a spin ball in real life) and select the distance of the ball from the batsman for the tip of the ball like in real life cricket. When bating, the user will decide either to strike or defend the ball. This is done through using random values to make the game feel more realistic.